nlp/requirements.txt
flask>=3.0.0
openai-whisper
ffmpeg-python
pydub
python-dotenv
requests
